+++ src/main/java/com/company/generatecheck/dto/DiscountCardDto.java
@@ -3,9 +3,11 @@ package com.company.generatecheck.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class DiscountCardDto {
 
-    private Double discount;
+    private BigDecimal discount;
 }
