+++ src/main/java/com/company/generatecheck/dto/ItemDto.java
@@ -3,12 +3,14 @@ package com.company.generatecheck.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ItemDto {
 
     private Integer qty;
     private ProductDto product;
-    private Double wholesaleDiscount;
-    private Double total;
+    private BigDecimal wholesaleDiscount;
+    private BigDecimal total;
 }
