+++ src/main/java/com/company/generatecheck/service/PriceService.java
@@ -3,13 +3,14 @@ package com.company.generatecheck.service;
 import com.company.generatecheck.model.DiscountCard;
 import com.company.generatecheck.model.Item;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 public interface PriceService {
 
-    Double getTaxablePrice(Set<Item> items);
+    BigDecimal getTaxablePrice(Set<Item> items);
 
-    Double getPromotionalDiscount(Double taxablePrice, DiscountCard discountCard);
+    BigDecimal getPromotionalDiscount(BigDecimal taxablePrice, DiscountCard discountCard);
 
-    Double getTotalPrice(Double priceWithTax, Double promotionalDiscount);
+    BigDecimal getTotalPrice(BigDecimal priceWithTax, BigDecimal promotionalDiscount);
 }
